implement player data

+++ src/main/java/pub/gdt/keke/data/NewNew.java
@@ -1,7 +1,7 @@
 package pub.gdt.keke.data;
 
 public interface NewNew {
-    void setLength(double length);
-    double getLength();
+    void setLength(int length);
+    int getLength();
     long getLastModificationTime();
 }
